[Front] Dashboard & Dataset Template (#66)
* feat: Add Footer component to layout and update page layout

The code changes in `layout.tsx` add a `Footer` component to the layout and update the page layout. This change improves the overall structure and design of the page.

The recent user commits and repository commits indicate that the developer has been working on adding new pages and improving the framework.

* feat: Add Footer component to layout and update page layout

* feat: Add Line and Liquid charts to Dashboard page

The code changes in `page.tsx` add Line and Liquid charts to the Dashboard page. This enhancement improves the visualization of traffic trend and usage data.

* feat: Add ProTable component to Dataset page

The code changes in `page.tsx` add the `ProTable` component to the Dataset page. This enhancement improves the functionality and display of the table, allowing users to interact with the data more efficiently.

* feat: Add ProTable component to Dataset page